advanced orbit generation

+++ apps/galaxy/renderables/Orbit.hpp
@@ -9,7 +9,7 @@ namespace galaxy {
 class OrbitRenderable : public pg::Renderable
 {
 public:
-    OrbitRenderable(const pgOrbit::OrbitalParameters<float>& op, int segments, pg::Color color)
+    OrbitRenderable(const pgOrbit::OrbitalParameters<double>& op, int segments, pg::Color color)
       : _color(color)
     {
         _points.reserve(segments + 1);
@@ -17,10 +17,10 @@ public:
         for (int i = 360; i-- > 0;)
         {
             // std constant for pi
-            auto p = pgOrbit::OrbitalMechanics<float>::getEulerAnglesFromEccentricAnomaly(
+            auto p = pgOrbit::OrbitalMechanics<double>::getEulerAnglesFromEccentricAnomaly(
                          op, i / 360.0 * 2 * std::numbers::pi)
                          .toCartesian();
-            _points.push_back({p[0], p[1]});
+            _points.push_back({static_cast<float>(p[0]), static_cast<float>(p[1])});
         }
 
         _points.push_back(_points[0]);
